Topologie, Doc, and glob
