+++ EMSL_api.py
@@ -7,6 +7,7 @@ Usage:
   EMSL_api.py list_basis        [--atom=<atom_name>...]
                                 [--db_path=<db_path>]
                                 [--average_mo_number]
+                                [--basis=<basis_name>...]
   EMSL_api.py list_atoms  --basis=<basis_name>
                                 [--db_path=<db_path>]
   EMSL_api.py get_basis_data --basis=<basis_name>
@@ -41,8 +42,8 @@ import sys
 import os
 
 from src.docopt import docopt
-from src.EMSL_utility import EMSL_dump
-from src.EMSL_utility import EMSL_local, checkSQLite3
+from src.EMSL_dump import EMSL_dump
+from src.EMSL_local import EMSL_local, checkSQLite3
 
 if __name__ == '__main__':
 
@@ -79,14 +80,16 @@ if __name__ == '__main__':
 
         amn = arguments["--average_mo_number"]
 
-        l = e.get_list_basis_available(elts, average_mo_number=amn)
+        l = e.get_list_basis_available(elts,
+                                       arguments["--basis"],
+                                       average_mo_number=amn)
 
         if amn:
             for name, des, avg in l:
-                print "{0} ({2}) | {1}".format(name, des, avg)
+                print "- '{}' ({}) || {:<50}".format(name, avg, des)
         else:
             for name, des in l:
-                print "{0} | {1}".format(name, des)
+                print "- {} || {:<50}".format(name, des)
 
     #  _     _     _     _____ _                           _
     # | |   (_)   | |   |  ___| |                         | |
